backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
biopython==1.81
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
Flask==2.3.3
Flask-Cors==4.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
MarkupSafe==3.0.2
numpy==1.26.4
openai==1.70.0
pydantic==2.11.1
pydantic_core==2.33.0
python-dotenv==1.0.0
requests==2.32.3
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
Werkzeug==2.3.7
zipp==3.21.0
# Updated RAG dependencies
langchain
langchain-community
langchain-ollama
langchain-chroma
langchain-huggingface
sentence-transformers==2.7.0
